Fix crash with Very Many Players mod
Add require = 0 to swapChannelClient and update the priority of mixin to 1049. The VMP's mixin has priority 1050 (inherited from vmp.mixins.json) and has require = 1 (inherited from defaultRequire = 1 at vmp.mixins.json) so it crashes if we keep the 6969 priority or require = 1 (inherited from mnet.mixins.json) in Modern Netty.

The swapChannelClient in Modern Netty seems to be only to cancel calling of awaitUninterruptibly and return the same channel back. This exactly what VMP does as well but it also sets a flag closing which is used to fix a deadlock.

All method are @Redirect in this mixin, except 1 @Inject, so priority does not cause unpredictable behaviour, there will be warnings on a conflict. As for the @Inject, changing the priority down might make some other mods @Injector's run before or after compared to before this commit, but since this injector is only a LOGGER.warn call in an exceptional scenario this is unimportant.

Alternative fix would be to remove defaultRequire from either mnet or vmp mixins.json, but that does result in mnet redirect going through causing is closing flag in VMP to not be set.

This the only proper fix I can think of.