Remove explicit priority from ServerConnectionListener mixin
All methods are @Redirect in this Mixin and the mod has defaultRequire = 1 which means if any of the redirectors conflict with another mod a warning and a game crash is issued automatically by the Mixin subsystem anyways, so this priority makes no sense. If in the future any methods are marked with explicit require = 0, this would also make the game not crash and only log the warning, which is the intent with an explicit require = 0. With an expliciit priority of such a high number (6969, nice) the game will crash even with require = 0 because other mods will have priority 1000 so mnet redirect would try to go through and crash.